+++ app/models/scan.rb
@@ -327,14 +327,14 @@ class Scan < ActiveRecord::Base
         save
 
         instance.service.suspend do |r|
-            next error_out if r.rpc_exception?
+            next error_out( r ) if r.rpc_exception?
 
             # Might take a while for the scan to finish suspending.
             probe_freq = 5
 
             Arachni::Reactor.global.delay( probe_freq ) do |task|
                 instance.service.snapshot_path do |path|
-                    next error_out if path.rpc_exception?
+                    next error_out( path ) if path.rpc_exception?
                     next Arachni::Reactor.global.delay( probe_freq, &task ) if !path
 
                     self.snapshot_path = path
@@ -438,7 +438,14 @@ class Scan < ActiveRecord::Base
             end
         end
 
-        instance.service.scan( options ) { refresh }
+        instance.service.scan( options ) do |r|
+            if r.rpc_exception?
+                error_out( r )
+                next
+            end
+
+            refresh
+        end
 
         self
     rescue => e
@@ -506,9 +513,11 @@ class Scan < ActiveRecord::Base
                             errors: error_messages.to_s.lines.count ],
                 without: [ issues: issue_digests ] ) do |progress_data|
 
+            self.error_messages ||= ''
+
             # ap progress_data
             if progress_data.rpc_exception?
-                error_out
+                error_out( progress_data )
                 next
             end
 
@@ -519,7 +528,7 @@ class Scan < ActiveRecord::Base
 
                 if progress_data[:errors] && progress_data[:errors].any?
                     self.error_messages ||= ''
-                    self.error_messages  += "\n#{progress_data[:errors].join( "\n" )}"
+                    self.error_messages  += "#{progress_data[:errors].join( "\n" )}\n"
                 end
                 save
 
@@ -645,7 +654,20 @@ class Scan < ActiveRecord::Base
 
     private
 
-    def error_out
+    def error_out( error = nil )
+        if error
+            self.error_messages ||= ''
+            self.error_messages  += "#{error}\n"
+
+            if error.respond_to? :backtrace
+                self.error_messages += "#{error.backtrace.join("\n")}\n"
+            end
+
+            self.error_messages += "#{'-' * 80}\n"
+            self.error_messages += "#{caller.join("\n")}\n"
+            self.error_messages += "#{'*' * 80}\n"
+        end
+
         self.status = :error
         finish